Start adding ninja based build system
Our most typical use case is the one where we know we have a
working compiler and we have all the dependencies already there
for cross compiling for mobile devices.

In such cases, wrestling with general purpose build systems is
a waste of time, and we can just use ninja to make what we want
with a little of bash-fu to generate a ninja.build.

The generated ninja.build also links measurement_kit. Even if
that binary is not going to be used, its existence is critical
because it allows us to spot an unsuccessful linking caused
by missing symbols. So we can catch more errors early on.