Refactor test/common to use less binaries
Each binary that we build, it means pulling in and compiling
the whole of Catch. If we just compile an object, instead, we
do not pull in the full complexity of Catch.

So, reducing to the minimum the number of test binaries should
cause the build to become faster.

Here there are just two binaries: one with everything and the
shared_ptr binary, which segfaults if included along with others
but works nicely if left alone (unclear to me why).